refactor: post request error redirects
